requirements.txt
azure-search-documents
openai
python-dotenv